+++ sqlparser/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.16.0-dev
 
 - New analysis checks for `RETURNING`: Disallow `table.*` syntax and aggregate expressions
+- Support `RAISE` expressions in triggers
 - Fix resolving columns when `RETURNING` is used in an `UPDATE FROM` statement
 - Fix aliases to rowid being reported as nullable